backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
requests
python-jose
httpx
pydantic-settings
cachetools
motor 
pymongo
